move test subject to a separate file
