+++ src/commands/test/test_results_queue.py
@@ -1,22 +1,10 @@
-from dataclasses import dataclass
-from pathlib import Path
-from typing import TYPE_CHECKING, List, Tuple
+from typing import TYPE_CHECKING, Tuple
 
 from src.commands.test.test_cases import TestCaseResult
-from src.utils.starknet_compilation import StarknetCompiler
+from src.commands.test.test_subject import TestSubject
 
 if TYPE_CHECKING:
     import queue
-
-
-@dataclass
-class TestSubject:
-    """
-    A dataclass consisting of identification of a single test bundle, and target functions
-    """
-
-    test_path: Path
-    test_functions: List[StarknetCompiler.AbiElement]
 
 
 class TestResultsQueue:
